nemometa: implement meta_transform function
Some element (such as those derived from gstbasetransform) will call the
meta_transform function to copy/transform buffer metadata into the new
buffer. Without this function implemented, the application will crash
as the call is done without checking if the meta implementation
implement this function.

I'm not sure if all meta implementations are expected to implement it or
not, but as this function is easy to implement in our case, we can just
implement it.